test/unit/: move test files to tests/ and stubDocument.js to setup-files/
