+++ test/unit/setup-files/stubDocument.js
